Split resize gesture and error boundary out of WidgetCard
The resize machinery (pointer capture, autoscroll loop, scroll-drift
compensation, clamping, motion values) moves to a useWidgetResize hook;
ErrorBoundary and WidgetRenderError move to WidgetErrorBoundary.tsx.
WidgetCard drops from 526 to 328 lines with no behavior change.
